pyproject.toml
[tool.poetry]
name = "rutilea-generative-defects-backend"
version = "0.1.0"
description = ""
authors = ["Mahdi Ghiasi <mahdiw.ghiasi@gmail.com>"]
readme = "README.md"
packages = [{include = "rutilea_generative_defects_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1.8"
djangorestframework = "^3.14.0"
django-filter = "^23.1"
django-cors-headers = "^3.14.0"
django-extensions = "^3.2.1"
drf-spectacular = "^0.26.1"
psycopg2 = "^2.9.6"
django-environ = "^0.10.0"
celery = "^5.2.7"
redis = "^4.5.4"
django-storages = "^1.13.2"
boto3 = "^1.26.112"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ipython = "^8.12.0"
django-debug-toolbar = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
